user authorization

+++ src/main/java/com/jinnova/smartpad/partner/PartnerManager.java
@@ -2,6 +2,7 @@ package com.jinnova.smartpad.partner;
 
 import java.sql.SQLException;
 import java.util.Iterator;
+import java.util.LinkedList;
 
 import com.jinnova.smartpad.db.BranchDao;
 import com.jinnova.smartpad.db.UserDao;
@@ -42,30 +43,57 @@ public class PartnerManager implements IPartnerManager {
 	}
 
 	@Override
-	public IUser createUser(IUser primaryUser, String login, String password) throws SQLException {
-		IUser u = new User(login, primaryUser.getBranchId());
+	public IUser createUser(IUser authorizedUser, String login, String password) throws SQLException {
+		if (!authorizedUser.isPrimary()) {
+			return null;
+		}
+		IUser u = new User(login, authorizedUser.getBranchId());
 		u.setPasshash(PartnerUtils.md5(password));
 		new UserDao().createUser(u);
 		return u;
 	}
 
 	@Override
-	public IUser loadUser(String login) throws SQLException {
-		return new UserDao().loadUser(login);
+	public IUser login(String login, String password) throws SQLException {
+		if (password == null || login == null) {
+			return null;
+		}
+		User u = (User) new UserDao().loadUser(login);
+		if (u == null) {
+			return null;
+		}
+		if (!PartnerUtils.md5(password).equals(u.getPasshash())) {
+			return null;
+		}
+		return u;
 	}
 
 	@Override
-	public void updateUser(IUser u) throws SQLException {
+	public void updateUser(IUser authorizedUser, IUser u) throws SQLException {
+		if (!authorizedUser.isPrimary() && !authorizedUser.getLogin().equals(u.getLogin())) {
+			return;
+		}
 		new UserDao().updateUser(u);
 	}
 
 	@Override
-	public void deleteUser(IUser u) throws SQLException {
+	public void deleteUser(IUser authorizedUser, IUser u) throws SQLException {
+		if (!authorizedUser.isPrimary()) {
+			return;
+		}
 		new UserDao().deleteUser(u);
 	}
 
 	@Override
-	public void updateBranch(IBranch branch) throws SQLException {
+	public LinkedList<IUser> listUsers(IUser authorizedUser) throws SQLException {
+		return new UserDao().listUsers(authorizedUser.getBranchId());
+	}
+
+	@Override
+	public void updateBranch(IUser authorizedUser, IBranch branch) throws SQLException {
+		if (!authorizedUser.isPrimary()) {
+			return;
+		}
 		new BranchDao().updateBranch(branch);
 	}
